requirements.txt
fastapi[standard]
uvicorn
google-generativeai
pydantic
pydantic-settings
python-dotenv
duckduckgo-search
slowapi
